requirements.txt
# 核心依赖
openai
python-dotenv
pyyaml
flask
flask-cors
aiohttp
matplotlib
numpy

# 生产环境依赖
gunicorn
gevent

python-pptx
pandas
scikit-learn
python-Levenshtein
nltk